backend/requirements.txt
fastapi
strawberry-graphql
uvicorn
python-multipart
Pillow

transformers==4.49.0
tokenizers==0.21.1
torch==2.5.1
sentencepiece==0.2.0
accelerate
datasets
openai
clip-openai
open-clip-torch
fairscale
taming-transformers
torch-fidelity

cairosvg
svgpathtools==1.6.1
svglib
reportlab
webcolors
lxml
seaborn==0.12.2
matplotlib
noise

httpx==0.24.0
pydantic==2.10
protobuf
omegaconf
scikit-learn==1.2.2
scikit-image
lpips
markdown2[all]


